Make the StudyLog test cases pass
